requirements.txt
absl-py==1.4.0
accelerate==1.2.1
addict==2.4.0
asttokens==3.0.0
astunparse==1.6.3
bitsandbytes==0.45.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
conda-pack==0.6.0
contourpy==1.0.7
cupy==13.5.1
cycler==0.11.0
diffusers==0.32.1
distlib==0.3.9
easydict==1.13
einops==0.6.0
exceptiongroup==1.2.2
executing==2.1.0
facexlib==0.3.0
fastrlock==0.8.3
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
flow-vis==0.1
fonttools==4.39.2
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.11.0
huggingface-hub==0.27.0
icecream==2.1.3
identify==2.6.3
idna==3.4
imageio==2.26.1
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.2
keras==3.5.0
kiwisolver==1.4.4
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.2
libclang==18.1.1
lightning-utilities==0.15.0
lit==15.0.7
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.3
lpips==0.1.4
Markdown==3.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.2.1
namex==0.0.8
natsort==8.4.0
networkx==3.0rc1
nodeenv==1.9.1
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==23.0
pandas==2.0.1
Pillow==9.3.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
protobuf==4.22.1
psutil==6.1.1
pwcnet==0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.18.0
pyiqa==0.1.13
pyparsing==3.0.9
pytest==8.3.4
python-dateutil==2.8.2
pytorch-triton==2.1.0+e650d3708b
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2024.11.6
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.20.0
scipy==1.10.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.3.21
timm==1.0.12
tokenizers==0.15.2
tomli==2.2.1
torch==2.5.1
torch-summary==1.4.5
torch-tb-profiler==0.4.3
torchaudio==2.1.0.dev20230321+cu117
torchmetrics==1.8.0
torchstat==0.0.7
torchvision==0.20.1
tqdm==4.65.0
transformers==4.37.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.13
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
yapf==0.43.0
zipp==3.21.0
